\name{seqDNAcopy-package}
\alias{seqDNAcopy-package}
\alias{seqDNAcopy}
\docType{package}
\title{
  \packageTitle{seqDNAcopy}
}
\description{
  \packageDescription{seqDNAcopy}
}
\details{
  seqDNAcopy implements the CBS algorithm for sequencing data. It uses
  coverage depth for tumor and normal data to compute log-ratio which is
  first normalized for GC content and then segmented using segment
  function from DNAcopy.
}
\author{
\packageAuthor{seqDNAcopy}

Maintainer: \packageMaintainer{seqDNAcopy}
}
\references{
  Need to be added
}
\keyword{ package }
